Product Edit Complete :)

+++ app/Http/Controllers/Admin/ProductController.php
@@ -7,9 +7,50 @@ use Request;
 use Redirect;
 use Input;
 use Session;
+use Auth;
+use Storage;
+use File;
 
 class ProductController extends BaseController
 {
+
+    public function deleteImage($id)
+    {
+        if (!Auth::check()) {
+            return Redirect::to('admin/auth/login');
+        }
+
+        // SET UPLOAD PATH
+        $destinationPath = 'resources/assets/images';
+        try {
+            $image = DB::table('images')
+                ->select('ID', 'Name')
+                ->where("ID", "=", $id)
+                ->where("RefTable", "=", "Products")
+                ->first();
+            if (count($image)) {
+                DB::beginTransaction();
+                $deleted = DB::table('images')->where('ID', '=', $id)->delete();
+                if ($deleted) {
+                    File::Delete($destinationPath . '/shirt.jpg');
+                }
+            }
+
+            return redirect()->back();
+        } catch (\Exception $e) {
+            DB::rollback();
+            $error = $e->getMessage();
+            if (env('Mode') == 'Development') {
+                $this->errorMsg = $error;
+                Session::flash('globalErrMsg', $this->errorMsg);
+                Session::flash('alert-class', 'alert-danger');
+            } else {
+                Session::flash('globalErrMsg', $this->errorMsg);
+                Session::flash('alert-class', 'alert-danger');
+            }
+            return redirect()->back();
+        }
+    }
 
     public function getProductCategories()
     {
@@ -52,6 +93,10 @@ class ProductController extends BaseController
 
     public function addProductView($productTypeID)
     {
+        if (!Auth::check()) {
+            return Redirect::to('admin/auth/login');
+        }
+
         $categories = DB::table('Categories')
             ->select('ID', 'Name')
             ->where("ProductTypeID", "=", 6)
@@ -91,8 +136,10 @@ class ProductController extends BaseController
 
     public function addProduct(Request $request)
     {
+        if (!Auth::check()) {
+            return Redirect::to('admin/auth/login');
+        }
 
-        //dd($array);
         $product = array();
 
         $product['ProductTypeID'] = $request::get('ProductTypeID');
@@ -355,9 +402,8 @@ class ProductController extends BaseController
             DB::table('ProductDetails')->insert($productDetailMain);
 
 
-            $return['status'] = true;
-            $return['msg'] = 'Product Added.';
-            $return['product'] = $pid;
+            Session::flash('globalSuccessMsg', 'Product Added.');
+            Session::flash('alert-class', 'alert-success');
 
             DB::commit();
 
@@ -377,11 +423,14 @@ class ProductController extends BaseController
 
         }
 
-
     }
 
     public function editProductView($id)
     {
+        if (!Auth::check()) {
+            return Redirect::to('admin/auth/login');
+        }
+
         if (empty($id)) {
             Session::flash('globalErrMsg', 'Product Not Found');
             Session::flash('alert-class', 'alert-danger');
@@ -429,6 +478,8 @@ class ProductController extends BaseController
                 unset($cat);
             }
         }
+
+        //dd($returnData);
         $selectedCategory = array();
         foreach ($product as $key => $value) {
             if ($value->PdRefTable == 'Categories') {
@@ -438,6 +489,48 @@ class ProductController extends BaseController
             }
         }
 
+        $relationTables = DB::table('CP')
+            ->select('CP.ID', 'CP.MPID', 'MP.ID as MPOID', 'MP.Name')
+            ->join('MP', 'CP.MPID', '=', 'MP.ID')
+            ->get();
+
+        $Opacity = '';
+        $Shine = '';
+        $WrinkleResistance = '';
+        $Thickness = '';
+        $Threadcount = '';
+        $Pattern = array();
+        $FabricType = array();
+        $CustomType = array();
+        $ShirtStyle = array();
+        $Season = array();
+        $SearchColors = array();
+        $ContrastColors = array();
+        foreach ($product as $key => $value) {
+            $oname = $getName = $this->getMPName($relationTables, 'ID', $value->PdRefID);
+            if (!empty($oname)) {
+                $getName = preg_replace('/\s+/', '', $getName);
+                if ($getName == 'Threadcount' || $getName == 'Thickness' || $getName == 'Opacity' || $getName == 'Shine' || $getName == 'WrinkleResistance') {
+                    ${$getName} = $value->PdRefID;
+                } else {
+                    if (is_array(${$getName}) && !in_array($value->PdRefID, ${$getName})) {
+                        array_push(${$getName}, $value->PdRefID);
+                        /*if($value->PdRefID == 170){
+                            $oname = $getName = $this->getMPName($relationTables, 'ID', $value->PdRefID);
+                            print_r(${$getName}); exit;
+                        }*/
+                    } else {
+                        ${$getName} = array();
+                        array_push(${$getName}, $value->PdRefID);
+                    }
+
+                    //echo $oname . ' Value is ' . print_r(${$getName}) . '<br />';
+                }
+
+            }
+
+        }
+
         //dd($product);
         $data = array(
             'categories' => $categories,
@@ -445,23 +538,33 @@ class ProductController extends BaseController
             'product' => $product[0],
             'productDetail' => $product,
             'productID' => $id,
-            'selectedCategory' => $selectedCategory
+            'selectedCategory' => $selectedCategory,
+            'Opacity' => $Opacity,
+            'Shine' => $Shine,
+            'WrinkleResistance' => $WrinkleResistance,
+            'Thickness' => $Thickness,
+            'Threadcount' => $Threadcount,
+            'Pattern' => $Pattern,
+            'FabricType' => $FabricType,
+            'CustomType' => $CustomType,
+            'ShirtStyle' => $ShirtStyle,
+            'Season' => $Season,
+            'SearchColors' => $SearchColors,
+            'ContrastColors' => $ContrastColors,
         );
-
         return view('admin/product/edit-product', $data);
     }
 
-    public function editProduct($id, Request $request)
+    public function editProduct(Request $request)
     {
-        $return = array(
-            'status' => false,
-            'msg' => ''
-        );
+        if (!Auth::check()) {
+            return Redirect::to('admin/auth/login');
+        }
 
         $product = array();
 
-        $product['ProductTypeID'] = $request::get('ProductTypeID');
-        $product['MerchantID'] = $request::get('MerchantID');
+        $ProductID = $request::get('productID');
+        $product['MerchantID'] = Session::get('userID');
         $product['Basic'] = $request::get('Basic');
         $product['Qty'] = $request::get('Qty');
 
@@ -480,25 +583,29 @@ class ProductController extends BaseController
         $product['EnableExpiry'] = ($request::get('EnableExpiry')) ? 1 : 0;
         $product['ExpiryDate'] = $request::get('EnableExpiryDate');
 
-        $product['Pos'] = $request::get('Pos');
-        $product['Status'] = $request::get('Status');
-        $product['MRID'] = $request::get('MRID');
-        $product['MRProductID'] = $request::get('MRProductID');
-
-        $product['Linked'] = $request::get('Linked');
-
         $productDetailMain = array();
-
         try {
             DB::beginTransaction();
-            DB::table('ProductDetails')->where('ProductID', '=', $id)->delete();
-            DB::table('Products')->where('ID', $id)->update($product);
+            DB::table('Products')->where('ID', $ProductID)->update($product);
+            DB::table('ProductDetails')->where('ProductID', '=', $ProductID)->delete();
+            if ($request::hasFile('picFile')) {
+                $files = Input::file('picFile');
+                $this->UploadImages($files, $ProductID, 0);
+            }
 
-            foreach ($request::except('Categories') as $category) {
+            if ($request::hasFile('zoomImg')) {
+                $zoomImg = Input::file('zoomImg');
+                $this->UploadImages($zoomImg, $ProductID, 1);
+            }
+            if ($request::has('Categories')) {
+                $array = $request::except('Categories');
+            }
+            unset($array['picFile']);
+            foreach ($array as $category) {
                 if (is_array($category) && count($category) > 0) {
                     foreach ($category as $cat) {
                         $productDetail = array();
-                        $productDetail['ProductID'] = $id;
+                        $productDetail['ProductID'] = $ProductID;
                         $productDetail['RefTable'] = 'CP';
                         $productDetail['RefID'] = $cat;
                         $productDetail['Extra'] = '';
@@ -511,10 +618,26 @@ class ProductController extends BaseController
                 }
             }
 
-            foreach ($request::get('Categories') as $category) {
+            if ($request::has('Categories')) {
+                foreach ($request::get('Categories') as $category) {
+                    $productDetail = array();
+                    $productDetail['ProductID'] = $ProductID;
+                    $productDetail['RefTable'] = 'Categories';
+                    $productDetail['RefID'] = $category;
+                    $productDetail['Extra'] = '';
+
+                    $productDetail['POS'] = 0;
+                    $productDetail['Def'] = '';
+
+                    array_push($productDetailMain, $productDetail);
+                }
+            }
+
+
+            /*foreach ($request::get('Categories') as $category) {
                 $productDetail = array();
-                $productDetail['ProductID'] = $id;
-                $productDetail['RefTable'] = 'Categories';
+                $productDetail['ProductID'] = $pid;
+                $productDetail['RefTable'] = 'CP';
                 $productDetail['RefID'] = $category;
                 $productDetail['Extra'] = '';
 
@@ -524,11 +647,11 @@ class ProductController extends BaseController
                 array_push($productDetailMain, $productDetail);
             }
 
-            if ((int)$request::get('Opacity') != 0) {
+            foreach ($request::get('ColorID') as $category) {
                 $productDetail = array();
-                $productDetail['ProductID'] = $id;
+                $productDetail['ProductID'] = $pid;
                 $productDetail['RefTable'] = 'CP';
-                $productDetail['RefID'] = (int)$request::get('Opacity');
+                $productDetail['RefID'] = $category;
                 $productDetail['Extra'] = '';
 
                 $productDetail['POS'] = 0;
@@ -537,11 +660,11 @@ class ProductController extends BaseController
                 array_push($productDetailMain, $productDetail);
             }
 
-            if ((int)$request::get('Thickness') != 0) {
+            foreach ($request::get('ContrastColorID') as $category) {
                 $productDetail = array();
-                $productDetail['ProductID'] = $id;
+                $productDetail['ProductID'] = $pid;
                 $productDetail['RefTable'] = 'CP';
-                $productDetail['RefID'] = (int)$request::get('Thickness');
+                $productDetail['RefID'] = $category;
                 $productDetail['Extra'] = '';
 
                 $productDetail['POS'] = 0;
@@ -550,11 +673,11 @@ class ProductController extends BaseController
                 array_push($productDetailMain, $productDetail);
             }
 
-            if ((int)$request::get('Threadcount') != 0) {
+            foreach ($request::get('CustomType') as $category) {
                 $productDetail = array();
-                $productDetail['ProductID'] = $id;
+                $productDetail['ProductID'] = $pid;
                 $productDetail['RefTable'] = 'CP';
-                $productDetail['RefID'] = $request::get('Threadcount');
+                $productDetail['RefID'] = $category;
                 $productDetail['Extra'] = '';
 
                 $productDetail['POS'] = 0;
@@ -563,11 +686,11 @@ class ProductController extends BaseController
                 array_push($productDetailMain, $productDetail);
             }
 
-            if ((int)$request::get('Shine') != 0) {
+            foreach ($request::get('FabricType') as $category) {
                 $productDetail = array();
-                $productDetail['ProductID'] = $id;
+                $productDetail['ProductID'] = $pid;
                 $productDetail['RefTable'] = 'CP';
-                $productDetail['RefID'] = (int)$request::get('Shine');
+                $productDetail['RefID'] = $category;
                 $productDetail['Extra'] = '';
 
                 $productDetail['POS'] = 0;
@@ -576,45 +699,154 @@ class ProductController extends BaseController
                 array_push($productDetailMain, $productDetail);
             }
 
-            if ((int)$request::get('WrinkleResistance') != 0) {
+            foreach ($request::get('Patterns') as $category) {
                 $productDetail = array();
-                $productDetail['ProductID'] = $id;
+                $productDetail['ProductID'] = $pid;
                 $productDetail['RefTable'] = 'CP';
-                $productDetail['RefID'] = (int)$request::get('WrinkleResistance');
+                $productDetail['RefID'] = $category;
                 $productDetail['Extra'] = '';
 
                 $productDetail['POS'] = 0;
                 $productDetail['Def'] = '';
 
                 array_push($productDetailMain, $productDetail);
+            }
+
+            foreach ($request::get('Season') as $category) {
+                $productDetail = array();
+                $productDetail['ProductID'] = $pid;
+                $productDetail['RefTable'] = 'CP';
+                $productDetail['RefID'] = $category;
+                $productDetail['Extra'] = '';
+
+                $productDetail['POS'] = 0;
+                $productDetail['Def'] = '';
+
+                array_push($productDetailMain, $productDetail);
+            }
+
+            foreach ($request::get('ShirtStyle') as $category) {
+                $productDetail = array();
+                $productDetail['ProductID'] = $pid;
+                $productDetail['RefTable'] = 'CP';
+                $productDetail['RefID'] = $category;
+                $productDetail['Extra'] = '';
+
+                $productDetail['POS'] = 0;
+                $productDetail['Def'] = '';
+
+                array_push($productDetailMain, $productDetail);
+            }*/
+            if ($request::has('Opacity')) {
+                if ((int)$request::get('Opacity') != 0) {
+                    $productDetail = array();
+                    $productDetail['ProductID'] = $ProductID;
+                    $productDetail['RefTable'] = 'CP';
+                    $productDetail['RefID'] = (int)$request::get('Opacity');
+                    $productDetail['Extra'] = '';
+
+                    $productDetail['POS'] = 0;
+                    $productDetail['Def'] = '';
+
+                    array_push($productDetailMain, $productDetail);
+                }
+            }
+
+            if ($request::has('Thickness')) {
+                if ((int)$request::get('Thickness') != 0) {
+                    $productDetail = array();
+                    $productDetail['ProductID'] = $ProductID;
+                    $productDetail['RefTable'] = 'CP';
+                    $productDetail['RefID'] = (int)$request::get('Thickness');
+                    $productDetail['Extra'] = '';
+
+                    $productDetail['POS'] = 0;
+                    $productDetail['Def'] = '';
+
+                    array_push($productDetailMain, $productDetail);
+                }
+            }
+
+            if ($request::has('Threadcount')) {
+                if ((int)$request::get('Threadcount') != 0) {
+                    $productDetail = array();
+                    $productDetail['ProductID'] = $ProductID;
+                    $productDetail['RefTable'] = 'CP';
+                    $productDetail['RefID'] = (int)$request::get('Threadcount');
+                    $productDetail['Extra'] = '';
+
+                    $productDetail['POS'] = 0;
+                    $productDetail['Def'] = '';
+
+                    array_push($productDetailMain, $productDetail);
+                }
+            }
+
+            if ($request::has('Shine')) {
+                if ((int)$request::get('Shine') != 0) {
+                    $productDetail = array();
+                    $productDetail['ProductID'] = $ProductID;
+                    $productDetail['RefTable'] = 'CP';
+                    $productDetail['RefID'] = (int)$request::get('Shine');
+                    $productDetail['Extra'] = '';
+
+                    $productDetail['POS'] = 0;
+                    $productDetail['Def'] = '';
+
+                    array_push($productDetailMain, $productDetail);
+                }
+            }
+
+            if ($request::has('WrinkleResistance')) {
+                if ((int)$request::get('WrinkleResistance') != 0) {
+                    $productDetail = array();
+                    $productDetail['ProductID'] = $ProductID;
+                    $productDetail['RefTable'] = 'CP';
+                    $productDetail['RefID'] = (int)$request::get('WrinkleResistance');
+                    $productDetail['Extra'] = '';
+
+                    $productDetail['POS'] = 0;
+                    $productDetail['Def'] = '';
+
+                    array_push($productDetailMain, $productDetail);
+                }
             }
 
 
             DB::table('ProductDetails')->insert($productDetailMain);
 
-            $return['status'] = true;
-            $return['msg'] = 'Product Updated.';
-            $return['product'] = $id;
-
             DB::commit();
-            echo json_encode($return);
+
+            Session::flash('globalSuccessMsg', 'Product Updated.');
+            Session::flash('alert-class', 'alert-success');
+            return Redirect::to('admin/product/products');
         } catch (\Exception $e) {
             DB::rollback();
             $error = $e->getMessage();
             if (env('Mode') == 'Development') {
                 $this->errorMsg = $error;
-                $return['msg'] = $this->errorMsg;
-
-                echo json_encode($return);
+                Session::flash('globalErrMsg', $this->errorMsg);
+                Session::flash('alert-class', 'alert-danger');
             } else {
-                $return['msg'] = $this->errorMsg;
-
-                echo json_encode($return);
+                Session::flash('globalErrMsg', $this->errorMsg);
+                Session::flash('alert-class', 'alert-danger');
             }
+            return redirect()->back()->withInput();
 
         }
+    }
 
-
+    public function getMPName($arr, $findCol, $findVal)
+    {
+        $find = false;
+        forEach ($arr as $val) {
+            if (!$find) {
+                if ($val->$findCol == $findVal) {
+                    $find = true;
+                    return $val->Name;
+                }
+            }
+        }
     }
 
     public function UploadImages($files, $productID, $zoom)
@@ -673,6 +905,9 @@ class ProductController extends BaseController
 
     public function getProducts()
     {
+        if (!Auth::check()) {
+            return Redirect::to('admin/auth/login');
+        }
         try {
 
             $products = DB::table('Products as pr')
@@ -745,6 +980,10 @@ class ProductController extends BaseController
 
     public function deleteProduct($id)
     {
+        if (!Auth::check()) {
+            return Redirect::to('admin/auth/login');
+        }
+
         try {
             DB::beginTransaction();
             DB::table('Products')->where('ID', '=', $id)->delete();
